Switch fcntl and ioctl to use zcall so that only libpizlo has to know about how to decode varargs. 0.7% regression on PizBench9000. Also teach ioctl to handle int args, so that the ncurses demo works reliably.

+++ filc/include/pizlonated_syscalls.h
@@ -35,7 +35,7 @@
    Currently, a lot of this is Linux-specific and the implementations assume that all structs are
    shaped according to their musl definitions. */
 
-int zsys_ioctl(int fd, unsigned long request, ...);
+int zsys_ioctl(int fd, int request, ...);
 long zsys_writev(int fd, const void* iov, int iovcnt);
 long zsys_read(int fd, void* buf, __SIZE_TYPE__ size);
 long zsys_readv(int fd, const void* iov, int iovcnt);
